web/app/app/requirements.txt
python-docx
flask-login
Flask-Principal
Flask-SQLAlchemy